backend/requirements.txt
python-dotenv
requests
aiogram